moved scripts to files, fixed bug in NPC script, Added A dialogue and Shogun dialogue
