--- Assets/Scripts/Camera_Shake.cs
@@ -1,13 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Camera_Shake : MonoBehaviour
-{
-    public Animator camAnim;
-
-    public void CamShake()
-    {
-        camAnim.SetTrigger("Shake");
-    }
-}
