--- Assets/Scripts/Camera_Shake.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1cea38b1c56d73c4c84b9831bdb4e298
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
