--- Assets/Scripts/CloudManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 88a33dae98f6e91468b0ade53d728131
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
