--- Assets/Scripts/Earthquake_Effect_Manager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ffcc25a1c9ec1844a8a83cbf7f2a4f99
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
