+++ Assets/Scripts/Enviroment.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 979d49a60e2745d588c76dbb230f20fc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
