--- Assets/Scripts/GameConstants.cs
@@ -1,6 +1,0 @@
-﻿public class GameConstants
-{
-    public static string k_ButtonNamePauseMenu = "Pause";
-    public static string k_ButtonNameSubmit = "Submit";
-    public static string k_ButtonNameCancel = "Cancel";
-}
