+++ Assets/Scripts/NPC/NPC.cs
@@ -32,13 +32,6 @@ public class NPC : MonoBehaviour
         foreach ( Interactable option in interactables ) option.enabled = false;
         activeInteractables = new bool[startQuestPhase.Length];
         EventManager.instance.onQuestProgressed += _check_interactable_options;
-        for ( int i =0; i < startQuestPhase.Length; i++ )
-        {
-            if ( startQuestPhase[i] == 0 )
-                activeInteractables[i] = true;
-            // if ( interactables[i].progressesQuestToPhase != -1 )
-            //     interactables[i].enabled = true;
-        }
     }
 
     void Update()
@@ -62,11 +55,30 @@ public class NPC : MonoBehaviour
             }
 
         }
-
         // If something is active, check if the interactables are disabled.
         else if (!interactables[activeOption].enabled)
             // if so, enable them.
             interactables[activeOption].enabled = true;
+        else //Check that if something is active that nothing of lower rank is now active 
+        {
+            int currActiveOption = activeOption;
+            int heighestRank = 99999;
+            // For all quests, determine...
+            for ( int i = 0; i < activeInteractables.Length; i++ )
+            {
+                // if the quest is active, and the rank of the interactables is less (higher priority?) than the highest rank.
+                if ( activeInteractables[i] && interactables[i].rank < heighestRank )
+                {
+                    // Set the new 'highest rank' (lowest priority?) to the corresponding interactable.
+                    heighestRank = interactables[i].rank;
+
+                    // Set the 'active quest' to that one.
+                    activeOption = i;
+                }
+            }
+            if ( currActiveOption != activeOption )
+                interactables[currActiveOption].enabled = false;
+        }
 
     }
 
@@ -78,7 +90,7 @@ public class NPC : MonoBehaviour
         // If there is an active queset, and this is that quest...
         if ( activeOption >= 0 && quest == quests[activeOption] )
         {
-            activeInteractables[activeOption] = false;              // disable it.
+            activeInteractables[activeOption] = false;       // disable it.
             interactables[activeOption].enabled = false;    // disable the corresponding interactables.
             activeOption = -1;                              // set the active quest to 'none'
         }
