+++ Assets/Scripts/Utility/GameConstants.cs
@@ -1,0 +1,6 @@
+﻿public class GameConstants
+{
+    public static string k_ButtonNamePauseMenu = "Pause";
+    public static string k_ButtonNameSubmit = "Submit";
+    public static string k_ButtonNameCancel = "Cancel";
+}
