+++ Assets/Scripts/Utility/GameConstants.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fa5ea6a0accb6f44ea25f5e8c62ea387
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
